+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## UNRELEASED
 ### Added
 ### Changed
+* ui: Show bin icon instead of minus if the quantity of the line item is 1
+
 ### Removed
 ### Fixed